+++ modules/item/weapon-item.js
@@ -60,10 +60,12 @@ export class HMWeaponItem extends HMItem {
         const misc      = {};
         const race      = {};
         const stats     = {};
+        const state     = {};
         const classData = actorData.bonus.class;
         const miscData  = actorData.bonus.misc;
         const statsData = actorData.bonus.stats;
         const raceData  = actorData.bonus.race;
+        const stateData = actorData.bonus.state;
 
         const spec      = {};
         const profTable = HMTABLES.weapons.noprof;
@@ -81,6 +83,7 @@ export class HMWeaponItem extends HMItem {
             misc[key]   = miscData?.[key] || 0;
             race[key]   = raceData?.[key] || 0;
             stats[key]  = statsData?.[key] || 0;
+            state[key]  = stateData?.[key] || 0;
 
             // Explicitly allowing multiple armor/shields because we don't support accesories yet.
             for (let i = 0; i < armors.length; i++)  {
@@ -105,6 +108,7 @@ export class HMWeaponItem extends HMItem {
         Object.values(armor).every((a) => a === 0)  ? delete bonus.armor  : bonus.armor  = armor;
         Object.values(shield).every((a) => a === 0) ? delete bonus.shield : bonus.shield = shield;
         Object.values(misc).every((a) => a === 0)   ? delete bonus.misc   : bonus.misc   = misc;
+        Object.values(state).every((a) => a === 0)   ? delete bonus.state : bonus.state  = state;
 
         if (isCharacter) {
             Object.values(stats).every((a) => a === 0)  ? delete bonus.stats : bonus.stats = stats;