+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+ - Functional Spell Fatigue status effect.
+ - Status effects auto apply/delete.
+
 ## [0.2.16] - 2022-09-05
 ### Added
 - Compendium items.